bot/pyproject.toml
[project]
name = "molten"
version = "0.0.0"
description = "Molten project automaton"
authors = [
    {name = "VincentRPS", email = "vincentbusiness55@gmail.com"},
]
dependencies = [
    "mafic>=2.9.3",
    "disnake[speed,voice]>=2.9.0",
    "redis>=5.0.1",
    "setuptools>=68.2.2",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "UNLICENSED"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.9.1",
    "isort>=5.12.0",
]

[tool.pdm.scripts]
start = { shell = "python -m molten" }

[tool.isort]
profile = "black"

[tool.black]
line-length = 120